html_scrapper/source_code/requirements.txt
requests==2.26.0
google-cloud-storage==1.43.0
requests-html==0.10.0
lxml==4.9.1
